Fix for #8
Had an error where root could not modify root-level services without
specifying the `--forceroot` option.

The fix involved skipping the root/local check if a root-level service
is being modified, as well as not creating root’s local TCC database if
it isn’t needed/if the option was not given.